improvements in environment control

+++ frontend/entrypoint.sh
@@ -2,11 +2,14 @@
 
 echo "Running entrypoint..."
 
-rm ./static -Rf
-npm run build
-npm run build-storybook
-mv ./build/ ./static
-mv ./storybook-static ./static/storybook
+if [ "$ENVIRONMENT" != "dev" ]; then
+  echo "Not in development mode! Building the static files..."
+  rm ./static -Rf
+  npm run build
+  npm run build-storybook
+  mv ./build/ ./static
+  mv ./storybook-static ./static/storybook
+fi;
 
 #Run container start command
 exec  "$@"
